+++ src/messages/lib/MessageTimeTravelDone.js
@@ -1,10 +1,11 @@
 const AbstractMessage = require('./AbstractMessage')
 
 class MessageTimeTravelDone extends AbstractMessage {
-	constructor ({ error }) {
+	constructor ({ error, timestamp }) {
 		super()
 		this.topic = 'time_travel_done'
 		this.error = error
+		this.timestamp = timestamp
 	}
 }